% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsl-package.R
\docType{package}
\name{phsl-package}
\alias{phsl}
\alias{phsl-package}
\title{phsl: Partitioned and Hybrid Methods for Learning Bayesian Network Structures}
\description{
Partitioned and hybrid methods for learning Bayesian network structures.
    This package implements the partitioned PC (pPC) algorithm, the p-value adjacency
    thresholding (PATH) algorithm, and the hybrid greedy initialization (HGI) algorithm
    for learning Bayesian network structures. Altogether, these algorithms form the 
    partitioned hybrid greedy search (pHGS) algorithm. These methods build upon and are 
    implemented so as to extend the capabilities of algorithms featured in the bnlearn 
    package.
}
\author{
\strong{Maintainer}: Jireh Huang \email{jirehhuang@ucla.edu}

Other contributors:
\itemize{
  \item Marco Scutari \email{marco.scutari@gmail.com} [contributor]
  \item Ariel Chow \email{achow12300@gmail.com} [reviewer]
}

}
\keyword{internal}
