% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsl-package.R
\docType{package}
\name{phsl-package}
\alias{phsl}
\alias{phsl-package}
\title{phsl: Partitioned Hybrid Methods for Learning Bayesian Network Structures}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jireh Huang \email{jirehhuang@ucla.edu}

Other contributors:
\itemize{
  \item Marco Scutari \email{marco.scutari@gmail.com} [contributor]
  \item Ariel Chow \email{arielcheng12@gmail.com} [reviewer]
}

}
\keyword{internal}
